# openapi.yml
openapi: 3.0.4
info:
  title: Acartia API
  version: "1.0"
servers:
  - url: /api/v1
paths:
...
  /auth/logout:
    post:
      summary: Sign out
      operationId: signOut
      responses:
        "200":
          description: Sign out successful
        "400":
          description: Not signed-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    ApiError:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
...